INT32 to Py_ssize_t
git-svn-id: http://192.168.20.11/sequoiadb/trunk@14572 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ driver/python/pysequoiadb/util.hpp
@@ -146,7 +146,7 @@
          goto done ;                                                    \
       }                                                                 \
                                                                         \
-      INT32 list_size = PyList_Size( py_list ) ;                        \
+      Py_ssize_t list_size = PyList_Size( py_list ) ;                   \
       for ( int idx = 0 ; idx < list_size ; ++idx )                     \
       {                                                                 \
          const bson::BSONObj *obj = NULL ;                              \
@@ -165,7 +165,7 @@
          goto done ;                                                    \
       }                                                                 \
                                                                         \
-      INT32 list_size = PyList_Size( py_list ) ;                        \
+      Py_ssize_t list_size = PyList_Size( py_list ) ;                   \
       for ( int idx = 0 ; idx < list_size ; ++idx )                     \
       {                                                                 \
          SINT64 id = PyLong_AsLongLong(PyList_GetItem( py_list, idx)) ; \
